+++ backend/rust/src/mwpm_decoder.rs
@@ -14,18 +14,18 @@ use std::sync::{Arc};
 pub struct MWPMDecoder {
     /// model graph is immutably shared
     pub model_graph: Arc<ModelGraph>,
-    /// complete model graph share precomputed data but each thread maintains variables
+    /// complete model graph each thread maintain its own precomputed data
     pub complete_model_graph: CompleteModelGraph,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct MWPMDecoderConfig {
-    /// default 0: no precomputed complete graph; set to positive integer to have finite entries for each graph vertex;
-    /// set to any negative number to have infinite entries for each graph vertex (may consume tons of time for large graph)
-    #[serde(alias = "pcmgms")]  // abbreviation
-    #[serde(default = "mwpm_default_configs::precompute_complete_model_graph_max_size")]
-    pub precompute_complete_model_graph_max_size: i64,
+    /// build complete model graph at first, but this will consume O(N^2) memory and increase initialization time,
+    /// disable this when you're simulating large code
+    #[serde(alias = "pcmg")]  // abbreviation
+    #[serde(default = "mwpm_default_configs::precompute_complete_model_graph")]
+    pub precompute_complete_model_graph: bool,
     /// weight function, by default using [`WeightFunction::AutotuneImproved`]
     #[serde(alias = "wf")]  // abbreviation
     #[serde(default = "mwpm_default_configs::weight_function")]
@@ -34,7 +34,7 @@ pub struct MWPMDecoderConfig {
 
 pub mod mwpm_default_configs {
     use super::*;
-    pub fn precompute_complete_model_graph_max_size() -> i64 { 0 }  // default to disable precomputed complete graph
+    pub fn precompute_complete_model_graph() -> bool { false }  // save for erasure error model and also large code distance
     pub fn weight_function() -> WeightFunction { WeightFunction::AutotuneImproved }
 }
 
@@ -47,12 +47,9 @@ impl MWPMDecoder {
         let mut simulator = simulator.clone();
         let mut model_graph = ModelGraph::new(&simulator);
         model_graph.build(&mut simulator, &error_model, &config.weight_function);
+        // build complete model graph
         let mut complete_model_graph = CompleteModelGraph::new(&simulator, &model_graph);
-        let precompute_complete_model_graph_max_size = if config.precompute_complete_model_graph_max_size >= 0 {
-            config.precompute_complete_model_graph_max_size as usize
-        } else { usize::MAX };
-        // compute complete model graph
-        complete_model_graph.precompute(&simulator, &model_graph, precompute_complete_model_graph_max_size);
+        complete_model_graph.precompute(&simulator, &model_graph, config.precompute_complete_model_graph);
         Self {
             model_graph: Arc::new(model_graph),
             complete_model_graph: complete_model_graph,