+++ CHANGELOG.md
@@ -10,7 +10,8 @@ For major updates, please run the [auto profiler](benchmark/auto_profiler/auto_p
 
 ## 0.1.4 (2022.3.30)
 
-- [ ] implement complete model graph using customized Dijkstra's algorithm
+- [x] build complete model graph using customized Dijkstra's algorithm
+- [ ] mwpm decoder
 
 ## 0.1.3 (2022.3.21)